+++ src/option_pricing/models/gbm/__init__.py
@@ -1,4 +1,5 @@
 from .simulation import GBMParams, simulate_gbm_paths, simulate_gbm_terminal
+from .simulators import GBMPathSimulator, GBMTerminalSimulator
 
 __all__ = [
     # params
@@ -6,4 +7,7 @@ __all__ = [
     # simulation functions
     "simulate_gbm_terminal",
     "simulate_gbm_paths",
+    # simulator adapters
+    "GBMTerminalSimulator",
+    "GBMPathSimulator",
 ]